+++ Editor/Scripts/Tools/ModifyTool.cs
@@ -4,7 +4,7 @@
 
 using System.Collections.Generic;
 using System.Linq;
-using InstancePainter;
+using InstancePainter.Runtime;
 using UnityEditor;
 using UnityEngine;